Refactoring: remove unnecessary unboxing

+++ main/src/cgeo/geocaching/connector/gc/GCParser.java
@@ -1502,7 +1502,7 @@ public abstract class GCParser {
                     final Integer ctl = Integer.valueOf(trackableMatcher.group(3));
                     final Integer id = Integer.valueOf(trackableMatcher.group(5));
                     if (trackCode != null && name != null && ctl != null && id != null) {
-                        final TrackableLog entry = new TrackableLog(trackCode, name, id.intValue(), ctl.intValue());
+                        final TrackableLog entry = new TrackableLog(trackCode, name, id, ctl);
 
                         Log.i("Trackable in inventory (#" + entry.ctl + "/" + entry.id + "): " + entry.trackCode + " - " + entry.name);
                         trackableLogs.add(entry);
