+++ tests/src/cgeo/test/Compare.java
@@ -42,7 +42,7 @@ public abstract class Compare {
                 assertTrue(actual.getAttributes().contains(attribute));
             }
             for (LogType logType : expected.getLogCounts().keySet()) {
-                assertTrue(actual.getLogCounts().get(logType).intValue() >= expected.getLogCounts().get(logType).intValue());
+                assertTrue(actual.getLogCounts().get(logType) >= expected.getLogCounts().get(logType));
             }
 
             // The inventories can differ too often, therefore we don't compare them. Also, the personal note
